feat(rss): Improve RSS feed
Feed got a very pretty style
(from https://github.com/genmon/aboutfeeds/blob/main/tools/pretty-feed-v3.xsl).

It is reverse date ordered, and slices to only display the last 50 posts.
Trailing slashes are removed from entries to have the same url as posts
themselves.

Fulltext is kept in the rss file - though I have to check how well it
works for mdx files and more complicated posts.
The content is parsed with markdown-it and sanitized with sanitize-html.

And it is made automatically discoverable on all pages by embedding into
the head of the base layout, so a feed reader should just automatically
find the rss file when pointed at any page.