feat(blog): Show conditional tags on blog posts
We now show tags on the top of a post, as well as on the bottom.

However, bottom tags are only shown for articles that have more than 600
words.

This seems a good length for the user having to scroll, so we avoid
'double-tags' on the screen if the article is short, but display bottom
tags for the user to continue surfing if the article is longer.