feat(repo): Update font selections
Since creating the site this is the first time I have started thinking
about fonts and font selections. So far it was very utilitarian -- just
use the default font given by tailwind and astro, or fall back to
whatever the browser recommends.

Now, we are purposefully selecting Roboto as the main page font and use
Lora for long-form text, with JetBrains mono as the code font.