fix(repo): Move to node lts debian container to fix builds
npm install has an issue with installing the dependencies on both
node:lts-alpine and node:lts-slim docker containers.

This changes it to simply be lts (bookworm debian) which does not seem
to have the same issues.
It may just be a missing git installation fixed with an `apk add git` or
`apt update && apt install git` but it seems more cumbersome than just
using the regular lts version for the build container instead.